extract python compat matrix support
